done: authentication feature
## fix
- database handler
> checking database exists.

## add
- GET sign-out in auth controller
> using jwt auth
- request.http
> http util on vscode
- __test__/auth.test.ts
> auth feature e2e test
- routes/auth.route.test.ts
> auth route unit test

+++ src/config/database.ts
@@ -53,7 +53,9 @@ class SequelizeSQLiteHandler implements IDatabaseHandler {
   async disconnect(): Promise<void> {
     await this._sequelize.close();
     if (this._env === 'test') {
-      fs.unlinkSync(this._dbPath);
+      if (fs.existsSync(this._dbPath)) {
+        fs.unlinkSync(this._dbPath);
+      }
     }
   }
 }
